Fixed a bunch of stuff
I added a loading screen before the level starts that works

I fixed the issue with menus

I made it so that that main menu and level select have a different game mode so that the timer doesn't work in there.

I started working on the single player mode. The switching between characters work, but it created a bunch of problems.