One Player Mode Update
Managed to make the UI work even in Single Player Mode.

There are still some bugs:
If I put a clean plate on the pan deposit, and then try to pick it up again, it will pick it up, but not visually, and so the plate teleports when putting it down on a different table.

Need to figure out a way to save high scores and how to separate the 2 game modes from each other.