Fix the character control bug, the level start cooldown bug, added basics for high score
I fixed the character controller bug by making it so that the spawns are only used in 2 player mode, and the characters placed in level only in 1 player mode.

The start level cooldown bug is a problem with Unreal and I fixed it by adding Event Tick to the Blueprint even if it's not connected.

Got a basis for a high-score system, but it doesn't work at the moment